+++ testutil/integration/common/grpc/bank.go
