+++ testutil/integration/evmos/factory/factory.go
@@ -7,24 +7,19 @@ import (
 	"fmt"
 	"math/big"
 
-	"github.com/cosmos/cosmos-sdk/client"
-	cosmostx "github.com/cosmos/cosmos-sdk/client/tx"
 	"github.com/cosmos/gogoproto/proto"
-
-	sdkmath "cosmossdk.io/math"
 
 	abcitypes "github.com/cometbft/cometbft/abci/types"
 	cryptotypes "github.com/cosmos/cosmos-sdk/crypto/types"
 	testutiltypes "github.com/cosmos/cosmos-sdk/types/module/testutil"
-	"github.com/cosmos/cosmos-sdk/types/tx/signing"
-	xauthsigning "github.com/cosmos/cosmos-sdk/x/auth/signing"
 
 	sdktypes "github.com/cosmos/cosmos-sdk/types"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/common/hexutil"
 	"github.com/ethereum/go-ethereum/crypto"
-	"github.com/evmos/evmos/v15/testutil/integration/grpc"
-	"github.com/evmos/evmos/v15/testutil/integration/network"
+	commonfactory "github.com/evmos/evmos/v15/testutil/integration/common/factory"
+	"github.com/evmos/evmos/v15/testutil/integration/evmos/grpc"
+	"github.com/evmos/evmos/v15/testutil/integration/evmos/network"
 	"github.com/evmos/evmos/v15/types"
 	evmtypes "github.com/evmos/evmos/v15/x/evm/types"
 
@@ -33,11 +28,9 @@ import (
 	"github.com/evmos/evmos/v15/server/config"
 )
 
-const (
-	GasAdjustment = float64(1.7)
-)
-
 type TxFactory interface {
+	commonfactory.TxFactory
+
 	// DeployContract deploys a contract with the provided private key,
 	// compiled contract data and constructor arguments
 	DeployContract(privKey cryptotypes.PrivKey, txArgs evmtypes.EvmTxArgs, deploymentData ContractDeploymentData) (common.Address, error)
@@ -46,8 +39,6 @@ type TxFactory interface {
 	// ExecuteEthTx builds, signs and broadcasts an Ethereum tx with the provided private key and txArgs.
 	// If the txArgs are not provided, they will be populated with default values or gas estimations.
 	ExecuteEthTx(privKey cryptotypes.PrivKey, txArgs evmtypes.EvmTxArgs) (abcitypes.ResponseDeliverTx, error)
-	// ExecuteEthTx builds, signs and broadcasts a Cosmos tx with the provided private key and txArgs
-	ExecuteCosmosTx(privKey cryptotypes.PrivKey, txArgs CosmosTxArgs) (abcitypes.ResponseDeliverTx, error)
 	// EstimateGasLimit estimates the gas limit for a tx with the provided address and txArgs
 	EstimateGasLimit(from *common.Address, txArgs *evmtypes.EvmTxArgs) (uint64, error)
 }
@@ -57,6 +48,7 @@ var _ TxFactory = (*IntegrationTxFactory)(nil)
 // IntegrationTxFactory is a helper struct to build and broadcast transactions
 // to the network on integration tests. This is to simulate the behavior of a real user.
 type IntegrationTxFactory struct {
+	*commonfactory.IntegrationTxFactory
 	grpcHandler grpc.Handler
 	network     network.Network
 	ec          *testutiltypes.TestEncodingConfig
@@ -69,9 +61,10 @@ func New(
 ) TxFactory {
 	ec := makeConfig(app.ModuleBasics)
 	return &IntegrationTxFactory{
-		grpcHandler: grpcHandler,
-		network:     network,
-		ec:          &ec,
+		IntegrationTxFactory: commonfactory.New(network, grpcHandler, &ec),
+		grpcHandler:          grpcHandler,
+		network:              network,
+		ec:                   &ec,
 	}
 }
 
@@ -151,21 +144,6 @@ func (tf *IntegrationTxFactory) ExecuteEthTx(
 	return res, nil
 }
 
-// ExecuteCosmosTx creates, signs and broadcasts a Cosmos transaction
-func (tf *IntegrationTxFactory) ExecuteCosmosTx(privKey cryptotypes.PrivKey, txArgs CosmosTxArgs) (abcitypes.ResponseDeliverTx, error) {
-	txBuilder, err := tf.buildTx(privKey, txArgs)
-	if err != nil {
-		return abcitypes.ResponseDeliverTx{}, errorsmod.Wrap(err, "failed to build tx")
-	}
-
-	txBytes, err := tf.encodeTx(txBuilder)
-	if err != nil {
-		return abcitypes.ResponseDeliverTx{}, errorsmod.Wrap(err, "failed to encode tx")
-	}
-
-	return tf.network.BroadcastTxSync(txBytes)
-}
-
 // EstimateGasLimit estimates the gas limit for a tx with the provided address and txArgs
 func (tf *IntegrationTxFactory) EstimateGasLimit(from *common.Address, txArgs *evmtypes.EvmTxArgs) (uint64, error) {
 	args, err := json.Marshal(evmtypes.TransactionArgs{
@@ -184,131 +162,6 @@ func (tf *IntegrationTxFactory) EstimateGasLimit(from *common.Address, txArgs *e
 	}
 	gas := res.Gas
 	return gas, nil
-}
-
-// buildTx builds a tx with the provided private key and txArgs
-func (tf *IntegrationTxFactory) buildTx(privKey cryptotypes.PrivKey, txArgs CosmosTxArgs) (client.TxBuilder, error) {
-	txConfig := tf.ec.TxConfig
-	txBuilder := txConfig.NewTxBuilder()
-
-	if err := txBuilder.SetMsgs(txArgs.Msgs...); err != nil {
-		return nil, errorsmod.Wrap(err, "failed to set tx msgs")
-	}
-
-	if txArgs.FeeGranter != nil {
-		txBuilder.SetFeeGranter(txArgs.FeeGranter)
-	}
-
-	senderAddress := sdktypes.AccAddress(privKey.PubKey().Address().Bytes())
-	account, err := tf.grpcHandler.GetAccount(senderAddress.String())
-	if err != nil {
-		return nil, errorsmod.Wrapf(err, "failed to get account: %s", senderAddress.String())
-	}
-
-	sequence := account.GetSequence()
-	signMode := txConfig.SignModeHandler().DefaultMode()
-	signerData := xauthsigning.SignerData{
-		ChainID:       tf.network.GetChainID(),
-		AccountNumber: account.GetAccountNumber(),
-		Sequence:      sequence,
-		Address:       senderAddress.String(),
-	}
-
-	// sign tx
-	sigsV2 := signing.SignatureV2{
-		PubKey: privKey.PubKey(),
-		Data: &signing.SingleSignatureData{
-			SignMode:  signMode,
-			Signature: nil,
-		},
-		Sequence: sequence,
-	}
-
-	err = txBuilder.SetSignatures(sigsV2)
-	if err != nil {
-		return nil, errorsmod.Wrap(err, "failed to set tx signatures")
-	}
-
-	if txArgs.FeeGranter != nil {
-		txBuilder.SetFeeGranter(txArgs.FeeGranter)
-	}
-
-	txBuilder.SetFeePayer(senderAddress)
-
-	gasLimit, err := tf.estimateGas(txArgs, txBuilder)
-	if err != nil {
-		return nil, errorsmod.Wrap(err, "failed to estimate gas")
-	}
-	txBuilder.SetGasLimit(gasLimit)
-
-	fees, err := tf.calculateFees(txArgs.GasPrice, gasLimit)
-	if err != nil {
-		return nil, errorsmod.Wrap(err, "failed to calculate fees")
-	}
-	txBuilder.SetFeeAmount(fees)
-
-	signature, err := cosmostx.SignWithPrivKey(signMode, signerData, txBuilder, privKey, txConfig, sequence)
-	if err != nil {
-		return nil, errorsmod.Wrap(err, "failed to sign tx")
-	}
-
-	err = txBuilder.SetSignatures(signature)
-	if err != nil {
-		return nil, errorsmod.Wrap(err, "failed to set tx signatures")
-	}
-
-	return txBuilder, nil
-}
-
-// encodeTx encodes the tx using the txConfig's encoder.
-func (tf *IntegrationTxFactory) encodeTx(txBuilder client.TxBuilder) ([]byte, error) {
-	txConfig := tf.ec.TxConfig
-	txBytes, err := txConfig.TxEncoder()(txBuilder.GetTx())
-	if err != nil {
-		return nil, errorsmod.Wrap(err, "failed to encode tx")
-	}
-	return txBytes, nil
-}
-
-// calculateFees calculates the fees for the transaction.
-func (tf *IntegrationTxFactory) calculateFees(gasPrice *sdkmath.Int, gasLimit uint64) (sdktypes.Coins, error) {
-	denom := tf.network.GetDenom()
-	var fees sdktypes.Coins
-	if gasPrice != nil {
-		fees = sdktypes.Coins{{Denom: denom, Amount: gasPrice.MulRaw(int64(gasLimit))}}
-	} else {
-		baseFee, err := tf.grpcHandler.GetBaseFee()
-		if err != nil {
-			return sdktypes.Coins{}, errorsmod.Wrap(err, "failed to get base fee")
-		}
-		price := baseFee.BaseFee
-		fees = sdktypes.Coins{{Denom: denom, Amount: price.MulRaw(int64(gasLimit))}}
-	}
-	return fees, nil
-}
-
-// estimateGas estimates the gas needed for the transaction.
-func (tf *IntegrationTxFactory) estimateGas(txArgs CosmosTxArgs, txBuilder client.TxBuilder) (uint64, error) {
-	txConfig := tf.ec.TxConfig
-	simulateBytes, err := txConfig.TxEncoder()(txBuilder.GetTx())
-	if err != nil {
-		return 0, errorsmod.Wrap(err, "failed to encode tx")
-	}
-
-	var gasLimit uint64
-	if txArgs.Gas == 0 {
-		simulateRes, err := tf.network.Simulate(simulateBytes)
-		if err != nil {
-			return 0, errorsmod.Wrap(err, "failed to simulate tx")
-		}
-
-		gasAdj := new(big.Float).SetFloat64(GasAdjustment)
-		gasUsed := new(big.Float).SetUint64(simulateRes.GasInfo.GasUsed)
-		gasLimit, _ = gasAdj.Mul(gasAdj, gasUsed).Uint64()
-	} else {
-		gasLimit = txArgs.Gas
-	}
-	return gasLimit, nil
 }
 
 // createMsgEthereumTx creates a new MsgEthereumTx with the provided arguments.
